requirements.txt
discord.py==1.7.3
discord-pretty-help==1.3.4
jishaku==2.5.0
googletrans==3.0.0